feat: adicionando verificacoes e autorizacao com o token JWT nas rotas de manipulacao das tarefas

+++ src/modules/Task/UseCase/DeleteTask/DeleteTaskUseCase.ts
@@ -9,6 +9,12 @@ class DeleteTaskUseCase {
     ) {}
 
     async execute(taskId: string, userId: string): Promise<void> {
+        const verificationTasExist = this.taskRepository.findTaskById(taskId);
+
+        if(!verificationTasExist) {
+            throw new Error('A task não foi encontrada no sistema!');
+        }
+
         this.taskRepository.deleteTaskById(taskId, userId);
     }
 }
